Updated the typings path
The typings when built now appear in the folder `dist/typings/*.d.ts` as opposed to `dist/typings/src/*.d.ts`. This appears to be a change from `0.0.1-alpha.19` to `1.0.0-beta.1`.

It's not obvious from the history exactly when this regressed but perhaps it could be somewhere within a Rollup plugin? (I'm not 100% familiar with Rollup or how the typings are generated).