chore: turn off Taobao yarn registry for Vue build
It looks like the reason Vue builds are randomly failing is because the
Vue CLI prompts the terminal to ask whether to use the Taobao registry
instead as the package install is taking too long. Of course, in a CI
build this prompt never gets answered, presumably times out and then we
never see it because we output the CLI command to `/dev/null`.

See https://cli.vuejs.org/guide/creating-a-project.html#vue-create