feat: support for hosted token worker (#1208)
<!-- By submitting a PR to this repository, you agree to the terms
within the [Auth0 Code of
Conduct](https://github.com/auth0/open-source-template/blob/master/CODE-OF-CONDUCT.md).
Please see the [contributing
guidelines](https://github.com/auth0/.github/blob/master/CONTRIBUTING.md)
for how to create and submit a high-quality PR for this repo. -->

### Changes

When the SDK is used in combination with a strict
Content-Security-Policy (CSP), the policy must include `worker-src:
blob:` which raises a concern of `unsafe-eval`. This change allows the
SDK to be configured to load the worker code from a trusted URL,
compliant with the CSP, and allows the user to mitigate the concern.

**Todo:**
- [x] Just waiting for https://github.com/auth0/auth0-spa-js/pull/1209
to land so that we can get e2e coverage of this

### References

> As defined above, special URL schemes that refer to specific pieces of
unique content, such as `data:`, `blob:` and `filesystem:` are excluded
from matching a policy of * and must be explicitly listed. Policy
authors should note that the content of such URLs is often derived from
a response body or execution in a Document context, which may be unsafe.
Especially for the
[default-src](https://www.w3.org/TR/CSP2/#default_src) and
[script-src](https://www.w3.org/TR/CSP2/#script_src) directives, policy
authors should be aware that allowing `data:` URLs is equivalent to
unsafe-inline and **allowing `blob:` or `filesystem:` URLs is equivalent
to unsafe-eval**.

https://www.w3.org/TR/CSP2/#source-list-guid-matching

### Testing

<!--
Please describe how this can be tested by reviewers. Be specific about
anything not tested and reasons why. If this library has unit and/or
integration testing, tests should be added for new functionality and
existing tests should complete without errors.
-->

- [X] This change adds unit test coverage
- [ ] This change adds integration test coverage
- [X] This change has been tested on the latest version of the
platform/language

### Checklist

- [X] I have read the [Auth0 general contribution
guidelines](https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md)
- [X] I have read the [Auth0 Code of
Conduct](https://github.com/auth0/open-source-template/blob/master/CODE-OF-CONDUCT.md)
- [X] All code quality tools/guidelines have been run/followed

---------

Co-authored-by: Frederik Prijck <frederik.prijck@auth0.com>