Update doc string explaining organization name parameter (#1333)
### Changes

A new security check is in place that requires Auth0 tenants to opt-in
when using the organization name as a parameter to the authorization
flow. This commit updates the doc string to reflect this change with a
link to the relevant documentation on the Auth0 website.

### References

I discussed this with a support staff member in ticket 02376965

You can also see another community member running into this issue here:

https://community.auth0.com/t/how-to-properly-sign-in-with-universal-login-and-organizations/150009

### Testing

Documentation update, no testing required

### Checklist

- [x] I have read the [Auth0 general contribution
guidelines](https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md)
- [x] I have read the [Auth0 Code of
Conduct](https://github.com/auth0/open-source-template/blob/master/CODE-OF-CONDUCT.md)
- [x] All code quality tools/guidelines have been run/followed

Co-authored-by: Ramya Anusri <62586490+ramya18101@users.noreply.github.com>