Fix outdated cache issue
If handleRedirectCallback is called without having previously called
logout, then it can result in outdated access tokens remaining in the
cache. This happens because handleRedirectCallback only updates the
cache with the token matching the audience/scope used in the login
request. If there are multiple tokens in the cache, only one of them
will be updated. This is a significant problem if logging in as a
new user, since there will now be tokens for multiple different users
in the cache with no way to tell them apart (at least not without
decoding them).

This issue is especially relevant for MRRT, as there are likely to be
several tokens in the cache with different audiences/scopes. It can,
however, also happen without MRRT if a different audience/scope was used
for login vs what was already in the cache.

To fix this, handleRedirectCallback will now clear the cache (the same
way logout does) before retrieving a new token.