[ESD-52771] Support mix of DPoP and Bearer in the fetcher (#1442)
### Changes

Applications can have a mix of APIs that use Bearer or DPoP, so the
fetcher needs to be able to make this decision based on the access token
type.

This is added in a non breaking way, when `getAccessToken` returns a
`string` - the behaviour remains the same (the token type is configured
by the `dpopNonceId` option)

### References

https://auth0team.atlassian.net/browse/ESD-52771

### Testing

<!--
Please describe how this can be tested by reviewers. Be specific about
anything not tested and reasons why. If this library has unit and/or
integration testing, tests should be added for new functionality and
existing tests should complete without errors.
-->

- [ ] This change adds unit test coverage
- [ ] This change adds integration test coverage
- [ ] This change has been tested on the latest version of the
platform/language

### Checklist

- [ ] I have read the [Auth0 general contribution
guidelines](https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md)
- [ ] I have read the [Auth0 Code of
Conduct](https://github.com/auth0/open-source-template/blob/master/CODE-OF-CONDUCT.md)
- [ ] All code quality tools/guidelines have been run/followed

---------

Co-authored-by: Frederik Prijck <frederik.prijck@okta.com>