Avoid saving token when MissingRefreshTokenError is thrown (#1448)
### Changes

We throw a `MissingRefreshTokenError` when the server, after a refresh,
responds with fewer scopes than the ones requested, and you don't have
`useRefreshTokensFallback` as true.

During the getTokenSilently flow, we create an entry in the cache, so
before the error is thrown, we should remove it to avoid storing the
token.

In this PR, it has been created a new method in the `cache-manager `to
accomplish that.

### References

Related to this issue: [When using MRRT, an error is thrown if any
requested scopes are missing in
response](https://github.com/auth0/auth0-spa-js/issues/1432)

### Testing

1. Initialize Auth0Client with `useMrrt` and `useRefreshTokens` to true:
```

const auth0 = new Auth0Client({
  domain: MY_DOMAIN,
  clientId: MY_CLIENT,
  useRefreshTokens: true,
  useMrrt: true,
  authorizationParams: {
    redirect_uri: REDIRECT_URI,
    audience: "audience1",
    scope: "read:orders",
  },
});
```
2. Log in with loginWithRedirect or loginWithPopup
3. Request a token for a different audience using getTokenSilently
```
const token = await auth0Client.getTokenSilently({
  authorizationParams: {
    audience: "audience2"
  },
});
```
- Note: `read:orders` scope does not exist in `audience2`

4. You will see that no token is saved with `audience2`.

- [ ] This change adds unit test coverage
- [ ] This change adds integration test coverage
- [x] This change has been tested on the latest version of the
platform/language

### Checklist

- [x] I have read the [Auth0 general contribution
guidelines](https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md)
- [x] I have read the [Auth0 Code of
Conduct](https://github.com/auth0/open-source-template/blob/master/CODE-OF-CONDUCT.md)
- [x] All code quality tools/guidelines have been run/followed

Co-authored-by: Gyanesh Gouraw <160731216+gyaneshgouraw-okta@users.noreply.github.com>